refactor: rename of step component to index
